+++ assets/svelte/consumers/FunctionPicker.svelte
@@ -19,15 +19,21 @@
   export let title: string = "Function";
   export let onFunctionChange: (functionId: string) => void;
   export let refreshFunctions: () => void;
-  export let functionTypes: string[] = ["function", "path"];
+  export let allowedFunctionTypes: string[] = ["function", "path"];
   export let showTypeLabel: boolean = true;
   export let typeLabelKey: string = "type";
   export let refreshState: "idle" | "refreshing" | "done" = "idle";
-  export let createNewQueryParams: string = "";
+  export let newFunctionType: string =
+    allowedFunctionTypes.length === 1 ? allowedFunctionTypes[0] : undefined;
+  export let routedSinkType: RoutedSinkType | undefined;
 
-  $: createNewLink = `/functions/new${createNewQueryParams}`;
-  $: filteredFunctions = functions.filter((t) =>
-    functionTypes.includes(t.type),
+  $: createNewLink = routedSinkType
+    ? `/functions/new?type=${newFunctionType}&sink_type=${routedSinkType}`
+    : `/functions/new?type=${newFunctionType}`;
+  $: filteredFunctions = functions.filter(
+    (t) =>
+      allowedFunctionTypes.includes(t.type) &&
+      (t.type !== "routing" || t.sink_type === routedSinkType),
   );
   $: hasNoneOption = $$slots["none-option"] !== undefined;
   $: hasValidOptions = filteredFunctions.length > 0 || hasNoneOption;