+++ lib/sequin/consumers/routing_function.ex
@@ -12,7 +12,14 @@ defmodule Sequin.Consumers.RoutingFunction do
   @primary_key false
   typed_embedded_schema do
     field :type, Ecto.Enum, values: [:routing], default: :routing
-    field :sink_type, Ecto.Enum, values: [:http_push, :redis_string, :nats]
+
+    field :sink_type, Ecto.Enum,
+      values: [
+        :http_push,
+        :redis_string,
+        :nats
+      ]
+
     field :code, :string
   end